新的深度学习测试用例，增加了初始化函数he，修改了初始化函数的参数类型，现在允许多种类型

+++ erud/opts/relu.py
@@ -13,5 +13,5 @@ class relu(payload) :
         _x = self.__x
 
         # 使用右导数，当xi为0时，dz/dxi = 1
-        dx = dz * (_x >= 0)
+        dx = dz * np.int64(_x >= 0)
         return [dx]
